third_party/WebKit/LayoutTests/external/wpt/fetch/api/request/request-clone.sub-expected.txt

This is a testharness.js-based test.
FAIL Check cloning a request assert_equals: Check cache attribute expected (string) "no-cache" but got (undefined) undefined
PASS Check cloning a request copies the headers 
Harness: the test ran to completion.
